+++ include/codegen_generic.h
@@ -4,9 +4,9 @@
 #ifndef _CODEGEN_H_
 #define _CODEGEN_H_
 
-#define MACHINE_REGISTER_SIZE_BYTES 8
-#define STACK_ALIGN_BYTES 16
-#define MAX_ASM_LINE_SIZE 256
+#define MACHINE_REGISTER_SIZE_BYTES ((size_t) 8)
+#define STACK_ALIGN_BYTES ((size_t) 16)
+#define MAX_ASM_LINE_SIZE ((size_t) 256)
 
 extern char printedLine[MAX_ASM_LINE_SIZE];
 extern char *registerNames[MACHINE_REGISTER_COUNT];
@@ -15,7 +15,7 @@ int ALIGNSIZE(unsigned int size);
 
 struct CodegenContext
 {
-    int *instructionIndex;
+    size_t *instructionIndex;
     FILE *outFile;
 };
 
@@ -27,7 +27,7 @@ void emitInstruction(struct TACLine *correspondingTACLine,
 char *PlaceLiteralStringInRegister(struct TACLine *correspondingTACLine,
                                    struct CodegenContext *context,
                                    char *literalStr,
-                                   int destReg);
+                                   u8 destReg);
 
 void verifyCodegenPrimitive(struct TACOperand *operand);
 
@@ -36,27 +36,27 @@ void WriteVariable(struct TACLine *correspondingTACLine,
                    struct Scope *scope,
                    struct LinkedList *lifetimes,
                    struct TACOperand *writtenTo,
-                   int sourceRegIndex);
+                   u8 sourceRegIndex);
 
 // places a variable in a register, with no guarantee that it is modifiable, returning the string of the register's name for asm
-int placeOrFindOperandInRegister(struct TACLine *correspondingTACLine,
+u8 placeOrFindOperandInRegister(struct TACLine *correspondingTACLine,
                                  struct CodegenContext *context,
                                  struct Scope *scope,
                                  struct LinkedList *lifetimes,
                                  struct TACOperand *operand,
-                                 int registerIndex);
+                                 u8 registerIndex);
 
-int pickWriteRegister(struct Scope *scope,
+u8 pickWriteRegister(struct Scope *scope,
                       struct LinkedList *lifetimes,
                       struct TACOperand *operand,
-                      int registerIndex);
+                      u8 registerIndex);
 
-int placeAddrOfLifetimeInReg(struct TACLine *correspondingTACLine,
+u8 placeAddrOfLifetimeInReg(struct TACLine *correspondingTACLine,
                              struct CodegenContext *context,
                              struct Scope *scope,
                              struct LinkedList *lifetimes,
                              struct TACOperand *operand,
-                             int registerIndex);
+                             u8 registerIndex);
 
 const char *SelectSignForLoad(u8 loadSize, struct Type *loaded);
 
@@ -69,47 +69,47 @@ char SelectWidthCharForLifetime(struct Scope *scope, struct Lifetime *lifetime);
 void EmitFrameStoreForSize(struct TACLine *correspondingTACLine,
                            struct CodegenContext *context,
                            enum riscvRegisters sourceReg,
-                           int size,
-                           int offset);
+                           u8 size,
+                           ssize_t offset);
 
 void EmitFrameLoadForSize(struct TACLine *correspondingTACLine,
                           struct CodegenContext *context,
                           enum riscvRegisters destReg,
-                          int size,
-                          int offset);
+                          u8 size,
+                          ssize_t offset);
 
 void EmitStackStoreForSize(struct TACLine *correspondingTACLine,
                            struct CodegenContext *context,
                            enum riscvRegisters sourceReg,
-                           int size,
-                           int offset);
+                           u8 size,
+                           ssize_t offset);
 
 void EmitStackLoadForSize(struct TACLine *correspondingTACLine,
                           struct CodegenContext *context,
                           enum riscvRegisters sourceReg,
-                          int size,
-                          int offset);
+                          u8 size,
+                          ssize_t offset);
 
 void EmitPushForOperand(struct TACLine *correspondingTACLine,
                         struct CodegenContext *context,
                         struct Scope *scope,
                         struct TACOperand *dataSource,
-                        int srcRegister);
+                        u8 srcRegister);
 
 void EmitPushForSize(struct TACLine *correspondingTACLine,
                      struct CodegenContext *context,
-                     int size,
-                     int srcRegister);
+                     u8 size,
+                     u8 srcRegister);
 
 void EmitPopForOperand(struct TACLine *correspondingTACLine,
                        struct CodegenContext *context,
                        struct Scope *scope,
                        struct TACOperand *dataDest,
-                       int destRegister);
+                       u8 destRegister);
 
 void EmitPopForSize(struct TACLine *correspondingTACLine,
                     struct CodegenContext *context,
-                    int size,
-                    int destRegister);
+                    u8 size,
+                    u8 destRegister);
 
 #endif
